Move `features/errors` to `components/errors`
