+++ scripts/tsc-with-whitelist.sh
@@ -7,8 +7,8 @@ web-admin/src/features/dashboards/listing/selectors.ts: error TS18048
 web-admin/src/features/dashboards/listing/selectors.ts: error TS2322
 web-admin/src/features/dashboards/listing/selectors.ts: error TS2345
 web-admin/src/features/dashboards/listing/selectors.ts: error TS2769
-web-admin/src/features/errors/error-utils.ts: error TS18048
-web-admin/src/features/errors/error-utils.ts: error TS2322
+web-admin/src/components/errors/error-utils.ts: error TS18048
+web-admin/src/components/errors/error-utils.ts: error TS2322
 web-admin/src/features/dashboards/query-mappers/mapQueryToDashboard.ts: error TS2322
 web-admin/src/features/dashboards/query-mappers/getDashboardFromAggregationRequest.ts: error TS2322
 web-admin/src/features/scheduled-reports/selectors.ts: error TS18048
