+++ web-admin/src/features/view-as-user/ViewAsUserPopover.svelte
@@ -12,7 +12,7 @@
     createAdminServiceSearchProjectUsers,
     type V1User,
   } from "../../client";
-  import { errorStore } from "../../features/errors/error-store";
+  import { errorStore } from "../../components/errors/error-store";
   import { viewAsUserStore } from "./viewAsUserStore";
 
   export let organization: string;
