+++ web-admin/src/routes/+error.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
   import { page } from "$app/stores";
   import ErrorPage from "@rilldata/web-common/components/ErrorPage.svelte";
-  import { createUserFacingError } from "../features/errors/error-utils";
+  import { createUserFacingError } from "../components/errors/user-facing-errors";
 
   $: ({ status, error } = $page); // Triggered by `throw error(status, message)`
